+++ app/src/main/java/woowacourse/paint/main/MainActivity.kt
@@ -4,6 +4,7 @@ import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
 import androidx.lifecycle.ViewModelProvider
 import woowacourse.paint.databinding.ActivityMainBinding
+import woowacourse.paint.main.adapter.ColorAdapter
 
 class MainActivity : AppCompatActivity() {
     private val binding by lazy { ActivityMainBinding.inflate(layoutInflater) }
